refector: fucntion 분리
